Remove networkmanager-cli from installed packages
Using nmcli from the SSH add-on is an unsupported way of adjusting
network configuration in HA OS, and the mismatch of versions of the
NetworkManager in the OS and nmcli in the add-on leads to errors that
even make it impossible to use it. Having nmcli bundled in the SSH
terminal add-on (even the unofficial one) makes users think this is
somehow the correct and supported way of configuring advanced
networking features.

Remove the package from the default set of installed software. Correct
way to handle network configuration in HA OS is to use `ha network`
command set, and if there's any functionality missing, open feature
requests for that.